Happy new year!
Restructuring Chapter 4 a little bit, flip flopping on my choice of cell complex vs triangulation in the post-subdivision world.

Subdivision and Cartesian product are simple to justify, but when handles start getting attached it would be best if there is no question by the reader as to whether the operation preserves structure as a manifold.

Will think it over.  Probably will subdivide cell complex into triangulation at end of Section 4.2.

Signed-off-by: Samuel Churchill <meadowschurchill@gmail.com>